requirements.txt
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-huggingface
pypdf
chromadb
python-dotenv
sentence-transformers
numpy